requirements.txt
google-api-python-client
pandas
numpy
matplotlib
requests
jupyter
opencv-python
pyiqa
scikit-image
torch
torchvision
tqdm
image-similarity-measures
pyfftw

